requirements.txt
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0
openpyxl==3.1.5
SQLAlchemy==2.0.30
scikit-learn==1.4.2
typer==0.12.3
gdown==5.2.0
tabulate==0.9.0
pytest==8.2.0
PyYAML==6.0.1